+++ lib/karafka/swarm/node.rb
@@ -27,20 +27,6 @@ module Karafka
       # @return [Integer] pid of the node
       attr_reader :pid
 
-      # When re-creating a producer in the fork, those are not attributes we want to inherit
-      # from the parent process because they are updated in the fork. If user wants to take those
-      # from the parent process, he should redefine them by overwriting the whole producer.
-      SKIPPABLE_NEW_PRODUCER_ATTRIBUTES = %i[
-        id
-        kafka
-        logger
-        oauth
-        polling
-        fd
-      ].freeze
-
-      private_constant :SKIPPABLE_NEW_PRODUCER_ATTRIBUTES
-
       # @param id [Integer] number of the fork. Used for uniqueness setup for group client ids and
       #   other stuff where we need to know a unique reference of the fork in regards to the rest
       #   of them.
@@ -72,34 +58,7 @@ module Karafka
           config.producer.close
 
           old_producer = config.producer
-          old_producer_config = old_producer.config
-
-          # Supervisor producer is closed, hence we need a new one here
-          config.producer = WaterDrop::Producer.new do |p_config|
-            p_config.kafka = Setup::AttributesMap.producer(kafka.dup)
-            p_config.logger = config.logger
-
-            old_producer_config.to_h.each do |key, value|
-              next if SKIPPABLE_NEW_PRODUCER_ATTRIBUTES.include?(key)
-
-              p_config.public_send("#{key}=", value)
-            end
-
-            # Namespaced attributes need to be migrated directly on their config node
-            old_producer_config.oauth.to_h.each do |key, value|
-              p_config.oauth.public_send("#{key}=", value)
-            end
-
-            old_producer_config.polling.to_h.each do |key, value|
-              next if SKIPPABLE_NEW_PRODUCER_ATTRIBUTES.include?(key)
-
-              p_config.polling.public_send("#{key}=", value)
-            end
-
-            old_producer_config.polling.fd.to_h.each do |key, value|
-              p_config.polling.fd.public_send("#{key}=", value)
-            end
-          end
+          config.producer = ProducerReplacer.new.call(old_producer, kafka, config.logger)
 
           @pid = ::Process.pid
           @reader.close