+++ spec/integrations/pro/cli/declaratives/health_multi_broker_sequential/single_broker_staging_spec.rb
@@ -44,8 +44,6 @@ draw_routes do
   end
 end
 
-create_routes_topics
-
 ARGV[0] = "topics"
 ARGV[1] = "health"