+++ spec/lib/karafka/admin/acl_spec.rb
@@ -2,7 +2,7 @@
 
 RSpec.describe_current do
   let(:resource_type) { :topic }
-  let(:resource_name) { "it-#{SecureRandom.uuid}" }
+  let(:resource_name) { generate_topic_name }
   let(:resource_pattern_type) { :literal }
   let(:principal) { "User:*" }
   let(:host) { "*" }